+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 
 ## 2.0.90
 
+#### Features
+* `EMP-11711` Playtoken passed as header when requesting fairplay licenses.
+
 #### Changes
 * Stop and unload `MediaAsset` when playback reaches end of duration.