+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-* `2.0.90` Release - [2.0.90](#2090)
+* `2.0.91` Release - [2.0.91](#2091)
 * `2.0.89` Release - [2.0.89](#2089)
 * `2.0.86` Release - [2.0.86](#2086)
 * `2.0.85` Release - [2.0.85](#2085)
@@ -10,7 +10,7 @@
 * `2.0.78` Release - [2.0.78](#2078)
 * `0.77.x` Releases - [0.77.0](#0770)
 
-## 2.0.90
+## 2.0.91
 
 #### Features
 * Added *Exposure* `X-Request-Id` to license requests.